+++ src/Components/Atoms/Images/index.js
@@ -1,6 +1,6 @@
 import React, { Fragment } from 'react';
 import {View,Text,StyleSheet,Image} from 'react-native';
-import styles from './style';
+//import styles from './style';
 
 const Images=(props)=>{
     const {photo} = props;
@@ -15,5 +15,13 @@ const Images=(props)=>{
 }
 
 
-
 export default Images;
+
+const styles=StyleSheet.create({
+    image:{
+        backgroundColor:'white',
+        width:230,
+        height:250,
+    },
+
+})
