+++ src/Components/Atoms/Images/style.js
@@ -2,7 +2,7 @@ import {StyleSheet} from 'react-native';
 
 const styles=StyleSheet.create({
     image:{
-        backgroundColor:'purple',
+        backgroundColor:'white',
         width:230,
         height:250,
     },
