+++ src/Components/Molecules/Note/index.js
@@ -1,14 +1,14 @@
 import React, { Fragment,useEffect } from 'react';
-import {View,Text,TouchableOpacity} from 'react-native';
+import {View,Text,TouchableOpacity,Image} from 'react-native';
 import styles from './style';
 import {getData} from './../../../hooks/firebase';
 import {Actions} from 'react-native-router-flux';
-
+import {Images} from './../../Atoms';
 const Note=(props)=>{
 
     const Edit=()=>{
         Actions.EditNote(props)
-        console.warn(props.titulo,props.contenido);
+        console.warn(props.titulo,props.contenido,props.imageUrl);
     }
 
     return(
@@ -20,7 +20,9 @@ const Note=(props)=>{
                         <Text style={styles.font}>{props.titulo}</Text>
                     </View>
                     <View style={styles.boxText}>
-                        <View style={styles.boxImage}></View>
+                        <View style={styles.boxImage}>
+                        <Images photo={props.imageUrl} style={styles.ImageStorage} />
+                        </View>
                         <View style={styles.text}>
                             <Text>{props.contenido}</Text>
                         </View>
