+++ src/Components/Molecules/Note/style.js
@@ -37,11 +37,16 @@ const styles=StyleSheet.create({
     },
     boxImage:{
        flex:1,
-        backgroundColor:'#F3D5F9',
+        backgroundColor:'yellow',
         margin:2,
         borderRadius:5,
         height:'80%',
+        width:'10%',
     },
+    ImageStorage:{
+         height:20,
+         width:20,
+     },
 
     text:{
         flex:6,
