+++ src/Components/Molecules/NoteBody/index.js
@@ -6,79 +6,13 @@ import Textarea from 'react-native-textarea';
 import firestore from '@react-native-firebase/firestore';
 import UploadPicker from './../../../../Upload';
 import {Input,Images}  from '../../Atoms'
-import ImagePicker from 'react-native-image-picker';
-import storage from '@react-native-firebase/storage';
 import * as Progress from 'react-native-progress';
 
 
-const NoteBody=({titulo,updateTitle,updateBody,contenido})=>{
-    const [imageUrl, setImageUrl] = useState(null);
-    const [image, setImage] = useState(null);
-    const [uploading, setUploading] = useState(false);
-    const [transferred, setTransferred] = useState(0);
+
+const NoteBody=({titulo,updateTitle,updateBody,contenido,imageUrl,image,uploadImage,selectImage,
+    uploading,transferred})=>{
     
-    const selectImage = () => {
-        const options = {
-          title:'Seleccionar imagen',
-          maxWidth: 2000,
-          maxHeight: 2000,
-          storageOptions: {
-            skipBackup: true,
-            path: 'images'
-          }
-        };
-        ImagePicker.showImagePicker(options, response => {
-          if (response.didCancel) {
-            console.log('User cancelled image picker');
-          } else if (response.error) {
-            console.log('ImagePicker Error: ', response.error);
-          } else if (response.customButton) {
-            console.log('User tapped custom button: ', response.customButton);
-          } else {
-            const source = { uri: response.uri };
-            console.log(source);
-            setImage(source);
-          }
-        });
-    };
-      const uploadImage = async () => {
-        const { uri } = image;
-        const filename = uri.substring(uri.lastIndexOf('/') + 1);
-        const uploadUri = Platform.OS === 'ios' ? uri.replace('file://', '') : uri;
-        setUploading(true);
-        setTransferred(0);
-        const task = storage()
-          .ref(filename)
-          .putFile(uploadUri);
-
-        // set progress state
-        task.on('state_changed', snapshot => {
-          setTransferred(
-            Math.round(snapshot.bytesTransferred / snapshot.totalBytes) * 10000
-          );
-        });
-        try {
-          await task;
-          let storageUrl = await storage()
-          .ref(filename).getDownloadURL();
-          setImageUrl(storageUrl);
-          console.warn(storageUrl);
-        } catch (e) {
-          console.error(e);
-        }
-        setUploading(false);
-        Alert.alert(
-          'Imagen guardada!',
-          'Imagen guardada en Firebase Cloud Storage.'
-        );
-       
-        setImage(null);
-        
-      
-    };
-
-
-
     return(
         <Fragment>
             <ScrollView style={styles.container}>
@@ -134,7 +68,7 @@ const NoteBody=({titulo,updateTitle,updateBody,contenido})=>{
           </View>
         </SafeAreaView>
 
-                <Images photo={imageUrl}/>
+                <Images photo={imageUrl} style={styles.imageBox7}/>
                 </View>
               </View>
             </ScrollView>
