+++ src/Components/Organisms/Content/index.js
@@ -14,7 +14,7 @@ const handlePress=()=>{
 
 const handleNote=({item})=>{
     //console.warn({item});
-    return <Note titulo={item.titulo} contenido={item.contenido} id={item.id}/>
+    return <Note titulo={item.titulo} contenido={item.contenido} id={item.id} imageUrl={item.imageUrl}/>
  }
